+++ src/org/francis/intel/challenge/stack/ByteStack.java
@@ -3,26 +3,26 @@ package org.francis.intel.challenge.stack;
 import java.util.Arrays;
 
 public class ByteStack {
-    private final byte[] stack;
+    private final int[] stack;
     private int idx = 0;
     
     public ByteStack(int size) {
-        stack = new byte[size];
+        stack = new int[size];
     }
     
-    public void push(byte pInt) {
+    public void push(int pInt) {
         stack[idx++] = pInt;
     }
     
-    public byte pop() {
+    public int pop() {
         return stack[--idx];
     }
     
-    public byte peek() {
+    public int peek() {
         return stack[idx-1];
     }
     
-    public byte peek(int lookback) {
+    public int peek(int lookback) {
         return stack[idx-(lookback+1)];
     }
     
@@ -32,7 +32,7 @@ public class ByteStack {
     
     @Override
     public String toString() {
-        byte[] print = new byte[idx];
+        int[] print = new int[idx];
         System.arraycopy(stack, 0, print, 0, idx);
         return Arrays.toString(print);
     }
